reorganized directories and added IdService
