+++ backend/src/main/java/com/github/bolabolor/backend/repository/FriendRepository.java
@@ -1,4 +1,5 @@
-package com.github.bolabolor.backend.friend;
+package com.github.bolabolor.backend.repository;
+import com.github.bolabolor.backend.model.Friend;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
