+++ backend/src/main/java/com/github/bolabolor/backend/repository/MongoUserRepository.java
@@ -1,4 +1,5 @@
-package com.github.bolabolor.backend.security;
+package com.github.bolabolor.backend.repository;
+import com.github.bolabolor.backend.model.MongoUser;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
