+++ backend/src/main/java/com/github/bolabolor/backend/service/CloudinaryService.java
@@ -1,4 +1,4 @@
-package com.github.bolabolor.backend.friend;
+package com.github.bolabolor.backend.service;
 import com.cloudinary.Cloudinary;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -10,9 +10,7 @@ import java.util.Map;
 
 @Service
 public class CloudinaryService {
-
     private final Cloudinary cloudinary;
-
     public CloudinaryService(Cloudinary cloudinary) {
         this.cloudinary = cloudinary;
     }
@@ -20,9 +18,7 @@ public class CloudinaryService {
     public String uploadImage(MultipartFile image) throws IOException {
         File fileToUpload = File.createTempFile("image", null);
         image.transferTo(fileToUpload);
-
         Map response = cloudinary.uploader().upload(fileToUpload, Collections.emptyMap());
-
         return response.get("url").toString();
     }
 }
