+++ backend/src/main/java/com/github/bolabolor/backend/service/IdService.java
@@ -1,0 +1,12 @@
+package com.github.bolabolor.backend.service;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import java.util.UUID;
+
+@Service
+@RequiredArgsConstructor
+public class IdService {
+    public String createRandomId() {
+        return UUID.randomUUID().toString();
+    }
+}
