+++ backend/src/test/java/com/github/bolabolor/backend/controller/UserControllerTest.java
@@ -1,4 +1,4 @@
-package com.github.bolabolor.backend;
+package com.github.bolabolor.backend.controller;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
