+++ backend/src/test/java/com/github/bolabolor/backend/service/FriendServiceTest.java
@@ -1,19 +1,13 @@
-package com.github.bolabolor.backend;
-
-import com.github.bolabolor.backend.friend.CloudinaryService;
-import com.github.bolabolor.backend.friend.Friend;
-import com.github.bolabolor.backend.friend.FriendRepository;
-import com.github.bolabolor.backend.friend.FriendService;
+package com.github.bolabolor.backend.service;
+import com.github.bolabolor.backend.model.Friend;
+import com.github.bolabolor.backend.repository.FriendRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.Collections;
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
+import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
@@ -24,7 +18,9 @@ class FriendServiceTest {
 
     FriendRepository friendRepository = mock(FriendRepository.class);
     CloudinaryService cloudinaryService = mock(CloudinaryService.class);
-    FriendService friendService = new FriendService(friendRepository, cloudinaryService);
+    final IdService idService = mock(IdService.class);
+    FriendService friendService = new FriendService(friendRepository, cloudinaryService, idService);
+
 
     @Test
     void getAllFriendsReturnEmptyList() {
@@ -54,20 +50,16 @@ class FriendServiceTest {
     @Test
     void addFriend() throws IOException {
         //GIVEN
-        Friend friendToSave = new Friend("123", "Air", null);
-
         MultipartFile file = Mockito.mock(MultipartFile.class);
+        Friend friend = new Friend( "Air");
+        Mockito.when(friendRepository.save(friend)).thenReturn(friend);
+        // WHEN
+        Friend actual = friendService.addFriend(friend, file);
 
-        when(cloudinaryService.uploadImage(any())).thenReturn("url");
-        when(friendRepository.save(friendToSave)).thenReturn(friendToSave);
-        //WHEN
-        Friend actual = friendService.addFriend(friendToSave, file);
         //THEN
-        Friend expected = new Friend("123", "Air", null);
-
-        verify(cloudinaryService).uploadImage(any());
-        verify(friendRepository).save(friendToSave);
-        assertEquals(expected, actual);
+        verify(friendRepository).save(friend);
+        verify(idService).createRandomId();
+        assertEquals(friend, actual);
     }
 
     @Test
